+++ packages/uni-app/uni_modules/mxuni-router-v2/js_sdk/plugins/index.d.ts
@@ -1,2 +1,2 @@
-export { A as AnimationPlugin, C as ChannelPlugin, I as InterceptorPlugin, o as ParamsPlugin, F as usePageChannel } from '../index-BShM2xTY.js';
+export { A as AnimationPlugin, C as ChannelPlugin, I as InterceptorPlugin, p as ParamsPlugin, H as usePageChannel } from '../index-R32Zb6OJ.js';
 import 'vue';